파일등록 완료, 상품 등록문제 확인 필요 20231024

+++ back/bootProject/src/main/java/com/bootProject/repository/File/FileRepositoryCoustom.java
@@ -4,6 +4,6 @@ import com.bootProject.entity.SaveFile;
 
 public interface FileRepositoryCoustom {
 
-    public int saveAllFile(SaveFile saveFile);
+    public long saveAllFile(SaveFile saveFile);
 
 }
